Refactor: relocate AbstractHierarchicalBuilder
