+++ src/main/java/de/fabiankeller/palladio/builder/repository/impl/SignatureBuilderImpl.java
@@ -1,5 +1,6 @@
 package de.fabiankeller.palladio.builder.repository.impl;
 
+import de.fabiankeller.palladio.builder.AbstractHierarchicalBuilder;
 import de.fabiankeller.palladio.builder.BuilderException;
 import de.fabiankeller.palladio.builder.repository.InterfaceBuilder;
 import de.fabiankeller.palladio.builder.repository.ParameterType;
