+++ src/main/java/de/fabiankeller/palladio/builder/resourceenvironment/impl/LinkBuilderImpl.java
@@ -1,6 +1,6 @@
 package de.fabiankeller.palladio.builder.resourceenvironment.impl;
 
-import de.fabiankeller.palladio.builder.repository.impl.AbstractHierarchicalBuilder;
+import de.fabiankeller.palladio.builder.AbstractHierarchicalBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.ContainerBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.LinkBuilder;
 import de.fabiankeller.palladio.builder.resourceenvironment.ResourceEnvironmentBuilder;
